Add the SQLiteAdapter
SQLiteAdapter is designed for the multi-profile Globus CLI case *and*
for higher level applications. It allows a single application to store
tokens with a naive "per-resource-server" scheme or to specify a
specialized scheme by putting data under customizable "namespaces" --
namespaces are partition keys for the data stored in a single DB for any
use-cases which want to leverage that functionality.

It defines everything in terms of stored JSON data so that the database
itself remains very simple.

Loading data out of the SQLiteAdapter is near-identical to loading a
by_resource_server dict out of a token response, and this is meant to
make integration easy.

The SQLiteAdapter also defines an auxiliary "config storage" which is
meant for arbitrary application data. Configs are just named json blobs
and this will be an important extension point for keeping data like
generated client credentials.

Tests cover the use of the sqlite adapter against a `:memory:` database
as well as multiple adapters sharing a single database file.